token_investigation/requirements.txt
nltk==3.2.1
numpy==1.11.2
pyenchant==1.6.8
